# api/openapi.yaml
openapi: 3.0.3
info:
  title: Pennywise API
  version: 1.0.0
  description: Self-hosted personal finance dashboard API
servers:
  - url: /api/v1
security:
  - cookieAuth: []
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List accounts for current user
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: Account list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTERNAL_ERROR
    PaginationMeta:
      type: object
      required:
        - page
        - per_page
        - total
        - total_pages
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    AccountType:
      type: string
      enum:
        - checking
        - savings
        - hysa
        - credit_card
        - mortgage
        - brokerage
        - retirement_401k
        - retirement_ira
        - retirement_roth_ira
        - rollover_ira
        - hsa
        - crypto_wallet
        - safe_agreement
        - credit_line
        - venmo
        - other
...
    Currency:
      type: string
      enum:
        - USD
        - BTC
        - ETH
        - EUR
...
    CreateAccountRequest:
      type: object
      required:
        - name
        - institution
        - account_type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        institution:
          type: string
          minLength: 1
          maxLength: 255
        account_type:
          $ref: "#/components/schemas/AccountType"
        currency:
          $ref: "#/components/schemas/Currency"
...
    AccountResponse:
      type: object
      required:
        - id
        - user_id
        - name
        - institution
        - account_type
        - currency
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        institution:
          type: string
        account_type:
          $ref: "#/components/schemas/AccountType"
        currency:
          $ref: "#/components/schemas/Currency"
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AccountListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AccountResponse"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
...